conversion fixes

+++ armsrc/string.h
@@ -17,7 +17,7 @@
 int strlen(const char *str);
 void *memcpy(void *dest, const void *src, int len);
 void *memmove(void *dest, const void *src, size_t len);
-void *memset(void *dest, int c, int len);
+void *memset(void *dest, uint8_t c, int len);
 int memcmp(const void *av, const void *bv, int len);
 void memxor(uint8_t *dest, uint8_t *src, size_t len);
 char *strncat(char *dest, const char *src, unsigned int n);
